[Sampler.BottomFloor] Target netstandard2.0 and net462
Follow the repository's library targeting convention: build for
$(TargetFrameworksForLibraries) plus the minimum .NET Framework on Windows,
so the package can be consumed from .NET Framework like the other components.

Reintroduce a downlevel path for the SDK's internal LogRecord.Copy(), which
UnsafeAccessor cannot reach below .NET 8. The fallback binds the method once
into a cached open delegate, so it costs no more per record than the accessor
and keeps the existing degradation behaviour when the method cannot be bound.

Replace the two APIs that do not exist downlevel: ArgumentNullException
.ThrowIfNull gives way to the repository's shared Guard helper, and
Random.Shared to a per-thread generator seeded from a Guid, since
new Random() on .NET Framework seeds from the tick count and would hand
threads started in the same tick identical sequences.

Run the tests on net462 as well, which is what exercises the reflection
fallback, and replace the test-only uses of double.IsFinite, Math.BitDecrement,
ToHashSet, GetValueOrDefault and range indexing so they compile there.

+++ test/OpenTelemetry.Sampler.BottomFloor.Tests/BottomFloorLogRecordProcessorTests.cs
@@ -55,7 +55,7 @@ public class BottomFloorLogRecordProcessorTests
         }
 
         Assert.Equal(budget, captured.Count);
-        Assert.All(captured, r => Assert.True(r.AdjustedCount > 0.0 && double.IsFinite(r.AdjustedCount)));
+        Assert.All(captured, r => Assert.True(r.AdjustedCount > 0.0 && Numeric.IsFinite(r.AdjustedCount)));
     }
 
     [Fact]
